pb array 3d to 2d

+++ fonctions.c
@@ -55,9 +55,9 @@ BOOL semestre_is_valid(Commande ma_commande){
     }
     return True;
 }
-int **create_2d_array(int row,int col){
-    static int **array=NULL;
-    array=malloc(sizeof(int*)*row);
+float **create_2d_array(int row,int col){
+    static float **array=NULL;
+    array=malloc(sizeof(float*)*row);
     if (array==NULL)
     {
         printf("pb alloc ");
@@ -65,7 +65,7 @@ int **create_2d_array(int row,int col){
     }
     for (size_t i = 0; i < row; i++)
     {
-        array[i]=malloc(sizeof(int)*col);
+        array[i]=malloc(sizeof(float)*col);
         if (array[i]==NULL)
         {
             printf("pb alloc ");
@@ -75,19 +75,19 @@ int **create_2d_array(int row,int col){
     return array;
     
 }
-void clear_2d_array(int **array,int row,int col){
+void clear_2d_array(float **array,int row,int col){
     for (size_t i = 0; i < row; i++)
     {
         free(array[i]);
     }
     free(array);
 }
-void print_array(int **array, int row , int col){
+void print_array(float **array, int row , int col){
     for (size_t i = 0; i < row; i++)
     {
         for (size_t j = 0; j < col; j++)
         {
-            printf("%d ",array[i][j]);
+            printf("%f ",array[i][j]);
         }
         printf("\n");
     }
